+++ src/App.tsx
@@ -22,7 +22,7 @@ export const App = () => {
               Home
             </NavLink>
 
-            <NavLink className={getLinkClass} to="people">
+            <NavLink className={getLinkClass} to="/people">
               People
             </NavLink>
           </div>
