+++ src/api.ts
@@ -1,6 +1,5 @@
 import { Person } from './types/Person';
 
-// eslint-disable-next-line operator-linebreak
 export const API_URL =
   'https://mate-academy.github.io/react_people-table/api/people.json';
