+++ src/components/Error/error.tsx
@@ -1,3 +1,5 @@
-export const Error = () => {
+import React from 'react';
+
+export const NotFoundPage = () => {
   return <h1 className="title">Page not found</h1>;
 };
