+++ src/components/People/people.tsx
@@ -2,20 +2,15 @@ import { Loader } from '../Loader/Loader';
 import { Person } from '../../types';
 import { useEffect, useState } from 'react';
 import { getPeople } from '../../api';
-import { PersonLink } from '../PersonLink/personLink';
 import '../../App.scss';
-import { useParams } from 'react-router-dom';
+
+import { PeopleTable } from '../PeopleTable/PeopleTable';
 
 export const People = () => {
   const [people, setPeople] = useState<Person[] | null>(null);
   const [loading, setLoading] = useState(false);
-  const { slug } = useParams();
-
-  const [, setPeopleLoadingError] = useState(false);
 
   const [errorMessage, setErrorMessage] = useState('');
-
-  const [selectedPersonSlug, setSelectedPersonSlug] = useState('');
 
   useEffect(() => {
     setErrorMessage('');
@@ -24,7 +19,6 @@ export const People = () => {
       .then(setPeople)
       .catch(() => {
         setErrorMessage('Something went wrong');
-        setPeopleLoadingError(true);
       })
       .finally(() => setLoading(false));
   }, []);
@@ -46,76 +40,7 @@ export const People = () => {
           )}
 
           {people && people.length > 0 && (
-            <table
-              data-cy="peopleTable"
-              className="table is-striped is-hoverable is-narrow is-fullwidth"
-            >
-              {!loading && (
-                <thead>
-                  <tr>
-                    <th>Name</th>
-                    <th>Sex</th>
-                    <th>Born</th>
-                    <th>Died</th>
-                    <th>Mother</th>
-                    <th>Father</th>
-                  </tr>
-                </thead>
-              )}
-              <tbody>
-                {people?.map(person => {
-                  const mother = people.find(p => p.name === person.motherName);
-                  const father = people.find(p => p.name === person.fatherName);
-
-                  return (
-                    <tr
-                      data-cy="person"
-                      key={person.slug}
-                      className={
-                        (
-                          slug
-                            ? person.slug === slug
-                            : person.slug === selectedPersonSlug
-                        )
-                          ? 'has-background-warning'
-                          : ''
-                      }
-                      onClick={() => setSelectedPersonSlug(person.slug)}
-                    >
-                      <td>
-                        <PersonLink person={person} />
-                      </td>
-
-                      <td>{person.sex}</td>
-                      <td>{person.born}</td>
-                      <td>{person.died}</td>
-                      <td>
-                        {person.motherName ? (
-                          mother ? (
-                            <PersonLink person={mother} />
-                          ) : (
-                            person.motherName
-                          )
-                        ) : (
-                          '-'
-                        )}
-                      </td>
-                      <td>
-                        {person.fatherName ? (
-                          father ? (
-                            <PersonLink person={father} />
-                          ) : (
-                            person.fatherName
-                          )
-                        ) : (
-                          '-'
-                        )}
-                      </td>
-                    </tr>
-                  );
-                })}
-              </tbody>
-            </table>
+            <PeopleTable people={people} loading={loading} />
           )}
         </div>
       </div>
